+++ src/core/attributerange.cpp
@@ -258,11 +258,6 @@ SingleValue::SingleValue(int id, const QString& attrName, Type type)
     m_max = Value("");
 
     switch (m_type) {
-    case Bool:
-        m_attrRangeStr = "bool";
-        m_min = Value(false);
-        m_max = Value(true);
-        break;
     case String:
         m_attrRangeStr = "string";
         break;
@@ -296,16 +291,45 @@ IntervalOfValues::IntervalOfValues(int id, const QString& attrName, Type type,
 
     switch (m_type) {
     case Bool:
+        Q_ASSERT(min.isBool() && max.isBool());
         m_attrRangeStr = "bool";
         f_rand = [](PRG* prg) { return prg->bernoulli(); };
+        f_next = [](const Value& v) { return v.type() == Value::BOOL ? !v.toBool() : v; };
+        f_prev = f_next;
         break;
     case Double_Range:
+        Q_ASSERT(min.isDouble() && max.isDouble());
         m_attrRangeStr = QString("double[%1,%2]").arg(min.toDouble()).arg(max.toDouble());
         f_rand = [this](PRG* prg) { return prg->uniform(m_min.toDouble(), m_max.toDouble()); };
+        f_next = [max, min](const Value& v) {
+            if (v.type() != Value::DOUBLE || v > max || v < min) return v;
+            if (v == max) return min;
+            double n = v.toDouble() + 1.0;
+            return n > max.toDouble() ? max : n;
+        };
+        f_prev = [max, min](const Value& v) {
+            if (v.type() != Value::DOUBLE || v > max || v < min) return v;
+            if (v == min) return max;
+            double n = v.toDouble() - 1.0;
+            return n < min.toDouble() ? min : n;
+        };
         break;
     case Int_Range:
+        Q_ASSERT(min.isInt() && max.isInt());
         m_attrRangeStr = QString("int[%1,%2]").arg(min.toInt()).arg(max.toInt());
         f_rand = [this](PRG* prg) { return prg->uniform(m_min.toInt(), m_max.toInt()); };
+        f_next = [max, min](const Value& v) {
+            if (v.type() != Value::INT || v > max || v < min) return v;
+            if (v == max) return min;
+            int n = v.toInt() + 1;
+            return n > max.toInt() ? min : n;
+        };
+        f_prev = [max, min](const Value& v) {
+            if (v.type() != Value::INT || v > max || v < min) return v;
+            if (v == min) return max;
+            int n = v.toInt() - 1;
+            return n < min.toInt() ? max : n;
+        };
         break;
     default:
         m_attrRangeStr.clear();
@@ -325,12 +349,15 @@ SetOfValues::SetOfValues(int id, const QString& attrName, Type type, Values valu
     m_attrRangeStr.clear();
     switch (m_type) {
     case Double_Set:
+        for (auto const& v : values) Q_ASSERT(v.type() == Value::DOUBLE);
         m_attrRangeStr = "double{";
         break;
     case Int_Set:
+        for (auto const& v : values) Q_ASSERT(v.type() == Value::INT);
         m_attrRangeStr = "int{";
         break;
     case String_Set:
+        for (auto const& v : values) Q_ASSERT(v.type() == Value::STRING);
         m_attrRangeStr = "string{";
         break;
     default:
@@ -343,6 +370,24 @@ SetOfValues::SetOfValues(int id, const QString& attrName, Type type, Values valu
         m_attrRangeStr += "," + values.at(i).toQString();
     }
     m_attrRangeStr += "}";
+}
+
+Value SetOfValues::next(const Value& val) const
+{
+    if (val.type() != m_min.type()) return val;
+    auto it = std::find(m_values.cbegin(), m_values.cend(), val);
+    if (it == m_values.cend()) return val; // val is not in the set
+    if (++it == m_values.cend()) return m_values.front(); // val is the last
+    return *it; // item after val
+}
+
+Value SetOfValues::prev(const Value& val) const
+{
+    if (val.type() != m_min.type()) return val;
+    auto it = std::find(m_values.cbegin(), m_values.cend(), val);
+    if (it == m_values.cend()) return val; // val is not in the set
+    if (it == m_values.cbegin()) return m_values.back(); // val is the first
+    return *(--it); // item before val
 }
 
 } // evoplex